fix the INT_MIN with casting && ternaries

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
  */
 int print_last_digit(int n)
 {
-	n = (n % 10 > 0) ? n % 10 : -n % 10;
+	n = (n % 10 > 0) ? n % 10 : -(long)n % 10;
 	_putchar(n + 48);
 	return (n);
 }
